relocate TestGetIndex to fix build tag conflict
